refactor: 예약된 작업을 Room에 저장하는 기능을 repository로 이동
